+++ app/src/main/java/com/techcourse/dao/ConnectionConsumer.java
@@ -1,4 +1,4 @@
-package com.techcourse.service;
+package com.techcourse.dao;
 
 import java.sql.Connection;
 import java.sql.SQLException;
